refactor(libcnb): group runtime resolvers
